+++ R/data.r
@@ -7,19 +7,19 @@
 #' Non-ASCII characters (e.g. those with accents) have been converted to ASCII 
 #' equivalents in this data frame.
 #'
-#' @format A data frame with eight variables and 15,251 rows:
-#' - `taxon_order`: numeric value used to sort rows in taxonomic order.
-#' - `category`: whether the entry is for a species or another
-#' field-identifiable taxon, such as `spuh`, `slash`, `hybrid`, etc.
-#' - `species_code`: a unique alphanumeric code identifying each species.
-#' - `common_name`: the common name of the species as used in eBird.
-#' - `scientific_name`: the scientific name of the species.
-#' - `order`: the scientific name of the order that the species belongs to.
-#' - `family`: the family of the species, in the form "Parulidae (New World
-#' Warblers)".
-#' - `report_as`: for taxa that can be resolved to true species (i.e. species,
-#' subspecies, and recognizable forms), this field links to the corresponding
-#' species code. For taxa that can't be resolved, this field is `NA`.
+#' @format A data frame with eight variables and 16,248 rows:
+#'   - `scientific_name`: scientific name.
+#'   - `common_name`: common name, defaults to English, but different languages 
+#'   can be selected using the `locale` parameter.
+#'   - `species_code`: a unique alphanumeric code identifying each species.
+#'   - `category`: whether the entry is for a species or another 
+#'   field-identifiable taxon, such as `spuh`, `slash`, `hybrid`, etc.
+#'   - `taxon_order`: numeric value used to sort rows in taxonomic order.
+#'   - `order`: the scientific name of the order that the species belongs to.
+#'   - `family`: the scientific name of the family that the species belongs to.
+#'   - `report_as`: for taxa that can be resolved to true species (i.e. species,
+#'   subspecies, and recognizable forms), this field links to the corresponding
+#'   species code. For taxa that can't be resolved, this field is `NA`.
 #'
 #' For further details, see \url{http://help.ebird.org/customer/en/portal/articles/1006825-the-ebird-taxonomy}
 #' @family data